Clarify time computations are done in POSIX time
Specifically, that they are done without leap seconds. This ensures
the issuance and expiration times of all batches are consistently
computed.

I've also moved 'Terminology and Roles' to 'Conventions and Definitions'
because it probably should go there.

(This is the same as what would happen if you implemented it naively,
but we need to be precise in a spec.)

Fixes #1