Only include hash of subject_info in Assertion
Move subject_info itself to BikeshedCertificate.

This reduces the amount of data that needs to be served by CAs a lot,
expecting large post-quantum public keys.

One downside is that it complicates the verification a bit: a verifier
should not forget to check whether subject_info_hash matches subject_info.

Also, to keep verification simpler, we make subject_info less flexible:
always expecting a hash in an assertion. We could go for a mix of hashed
and unhashed subject info in the assertion, but I do not see an
application yet.