Rework discussion of the transparency ecosystem
Although the transparency service was clearly described throughout the
document as an abstraction and not actually necessarily one single
centralized service, this seems to have been a point of confusion. The
intent was to observe that this system actually allows the ecosystem to
freely design trust models, mirroring roles, etc., all without impact to
interoperability. As this all happens out of band, neither certificate
issuance nor certificate usage is impacted by these choices, except that
a certificate may or may not be selected by negotiation.

This seems to have been a confusing notion. To avoid that mess, recut
these concepts in the document.

First, define the role of a mirror as a concrete service with some
responsibility. This picks up the mirroring aspects of the old example
transparency services, but not the direct connection to an RP or
trustworthiness. It's a freefloating service with some responsibility.
You may or may not believe it meets those responsibilities.

Then, define a relying party policy, for how an RP gets gets validity
windows. This is more explicitly the source of implementation
dependence. RPs can choose to trust some combination of mirrors when
formulating this policy. We give several example policies, but don't
mandate anything in particular.

Rewrite the rest of this text according to this framing.

Closes #99.