Update citations
draft-beck-tls-trust-anchor-ids is not draft-ietf-tls-trust-anchor-ids

We ended up with a mix of I-D.draft-foo and I-D.foo. It seems they both
work but result in different citations. Align on the latter because it's
shorter. (Also I vaguely recall, in other drafts, that it was the
canonical one, but I could be completely misremembering.)