Add types for most artwork/sourcecode blocks
xml2rfc v3 distinguishes between <artwork> and <sourcecode> now. It
seems kramdown-rfc handles it by treating untyped blocks as <artwork>,
and almost all types as <sourcecode>, except for the handful of ones set
for artwork.

https://www.rfc-editor.org/rpc/wiki/doku.php?id=sourcecode-types has
some "recommended" values for <sourcecode> types, so I've filled tehm
in. The only non-obvious case was the example X.500 name. I've left that
alone.

This seems to have no effect on the rendering, just the resulting XML.