Rename 'landmark checkpoints' to 'landmarks'
"Landmarks" serve a different role than "checkpoints", so rename to
reflect that.  In tlog-tiles/static-ct-api, "checkpoints" are signed
tree heads/root hashes to summarize the state of the tree. "Landmarks",
on the other hand, represent tree sizes (not even tree heads) and trust
anchor IDs and are used for coordination amongst parties in the
ecosystem.  (As a bonus, "landmark sequence" is less of a mouthful than
"landmark checkpoint sequence").

Clarify that landmarks don't necessarily need to be aligned with
checkpoint tree sizes, and technically could be allocated at any tree
sizes less or equal to the latest checkpoint tree size (although not
sure why you'd do this).

Also clarify that unexpired entries don't necessarily need to be covered
by a landmark subtree, as they can be between the last landmark subtree
and the log's latest checkpoint.