Clarify verification process
Matt Mueller noticed that the process was ambiguous about where the
MTCProof fields came from, and that we didn't consistently rename the
OID to match the id-alg-* convention that LAMPS seems to be using these
days.

From there I also noticed that the shape of RFC 9162, section 2.1.3.2
doesn't quite fit what we need. We want to run the inclusion proof, get
the expected subtree hash, and then check a signature over it. RFC 9162
assumes you already know the subtree hash and are just checking it. I
copy-pasted the description from RFC 9162 and inlined the small changes
we were making.