Unify the two branches in the consistency proof check
Also replace prepending to the proof list with removing in the other
branch. It makes it a little obvious how to implement this without
making a copy of the proof. This should also hopefully make the
relationship between inclusion and consistency proof evaluation clearer.

While I'm here, use '1.' for the steps so we don't have to renumber it
as it changes.