Restore the more efficient trust anchor negotiation
When trust_anchors was extracted from the first MTC drafts, we lost an
optimization because it was not, at the time, relevant for the baseline
version of the problem. However, it is still quite useful for MTCs (and
may be useful for other scenarios too).

This implements https://github.com/tlswg/tls-trust-anchor-ids/issues/62
to put it back. For now, it's in the MTC repo, but in the long run, we
should put it back in TAI itself.